% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{apriori.hydropressure.difference.samples}
\alias{apriori.hydropressure.difference.samples}
\title{Apiori hydrostatic pressure samples based on time-delta}
\usage{
apriori.hydropressure.difference.samples(interval.sec)
}
\description{
This function extracts samples from apriori hydrostatic pressure data.
These samples can then be used as estimates of \eqn{\epsilon(\Delta t)}.
}
\keyword{internal}
