% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apriori.R
\name{apriori}
\alias{apriori}
\title{A-priori data information class}
\usage{
apriori(data_type = c("air pressure", "diver"), units = c("cmH2O"))
}
\arguments{
\item{data_type}{character, specifying type of data}

\item{units}{character, specifying units of data}
}
\value{
A-priori object.
}
\description{
This function creates an apriori class based on input.
}
\examples{
# For example:
ap <- apriori("air pressure", "cmH2O")

}
