% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation.R
\name{compensate.altitude}
\alias{compensate.altitude}
\title{Air pressure altitude compensation}
\usage{
compensate.altitude(x, altitude.m)
}
\arguments{
\item{altitude.m}{the exact altitude at which the barometer operates.}
}
\value{
x reduced to MSL pressure.
}
\description{
This function is for air pressure altitude compensation based on the barometric formula.
}
\keyword{internal}
