% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\docType{data}
\name{CONST.ALPHA}
\alias{CONST.ALPHA}
\title{Type I error rate constant}
\format{An object of class \code{numeric} of length 1.}
\usage{
CONST.ALPHA
}
\description{
Type I error rate constant
}
\keyword{internal}
