% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{decay}
\alias{decay}
\title{Decay function for x}
\usage{
decay(index, decay, n)
}
\description{
Decay function for TC exponential decay when the model is written
in function of x only.
}
\keyword{internal}
