% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{detect}
\alias{detect}
\title{Hydrostatic pressure event detection function}
\usage{
detect(x, timestamps, types = c("AO", "LS", "TC"), nr.tail = 25)
}
\value{
Returns an \link{Events} object.
}
\description{
This function will take the hydrostatic pressure data x, and the
corresponding timestamps, and will use \link{seeker} function to select the
best model within a window. It will repreat until all x is processed.
}
\keyword{internal}
