% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drifts.R
\name{detect_drift}
\alias{detect_drift}
\alias{detect_drift,numeric,apriori-method}
\title{Detect drift}
\usage{
detect_drift(
  x,
  timestamps,
  apriori = apriori("air pressure", units = "cmH2O"),
  ...,
  plot = FALSE,
  verbose = FALSE,
  title = NULL
)

\S4method{detect_drift}{numeric,apriori}(x, timestamps, apriori, plot, verbose, title)
}
\arguments{
\item{x}{numeric vector of values}

\item{timestamps}{timestamp vector}

\item{apriori}{\link{apriori} class. Defaults to air pressure.}

\item{...}{optional parameters, depending on signature:}

\item{plot}{prints comprehensive plots}

\item{verbose}{prints comprehensive information}

\item{title}{adds title to the plot}
}
\value{
Logical vector with same length as x, specifying TRUE for drifting observations.
}
\description{
This function marks a drift in the input vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = numeric,apriori = apriori}: 
}}

\references{
\href{https://dov-vlaanderen.github.io/groundwater-logger-validation/gwloggeR/docs/articles/Airpressure.html}{Air pressure vignette}
}
