% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{detect_duplicates}
\alias{detect_duplicates}
\title{Detect duplicate timestamps}
\usage{
detect_duplicates(ts)
}
\arguments{
\item{ts}{timestamp vector of type Date or POSIXct}
}
\value{
Logical vector with same length as x, specifying TRUE for a duplicate value.
}
\description{
This function marks duplicate timestamps in the input vector.
}
\examples{
# In case of a vector:
ts <- seq(as.Date('2019-01-01'), as.Date('2019-01-5'), by = 'day')
ts <- c(ts, as.Date('2019-01-5'))
detect_duplicates(ts)

}
