% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelshifts.R
\docType{methods}
\name{detect_levelshifts}
\alias{detect_levelshifts}
\alias{detect_levelshifts,numeric,apriori-method}
\title{Detect levelshifts}
\usage{
detect_levelshifts(x, timestamps, apriori, ..., plot = FALSE,
  verbose = FALSE, title = NULL)

\S4method{detect_levelshifts}{numeric,apriori}(x, timestamps, apriori,
  plot, verbose, title)
}
\arguments{
\item{x}{numeric vector of values}

\item{timestamps}{timestamp vector}

\item{apriori}{\link{apriori} class}

\item{...}{optional parameters, depending on signature:}

\item{plot}{prints comprehensive plots}

\item{verbose}{prints comprehensive information}

\item{title}{adds title to the plot}
}
\value{
Logical vector with same length as x, specifying TRUE for an levelshift start.
}
\description{
This function marks levelshifts in the input vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = numeric,apriori = apriori}: 
}}

