% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_norm.R
\name{detect_outliers_norm}
\alias{detect_outliers_norm}
\title{Detects outliers}
\usage{
detect_outliers_norm(x, alpha = 5e-04, verbose = FALSE, x.mean, x.sd,
  type = c("two.sided", "less", "greater"))
}
\description{
...
}
\keyword{internal}
