% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalchanges.R
\docType{methods}
\name{detect_temporalchanges}
\alias{detect_temporalchanges}
\alias{detect_temporalchanges,numeric,apriori-method}
\title{Detect Temporal changes}
\usage{
detect_temporalchanges(x, timestamps, apriori, ..., plot = FALSE,
  verbose = FALSE, title = NULL)

\S4method{detect_temporalchanges}{numeric,apriori}(x, timestamps, apriori,
  plot, verbose, title)
}
\arguments{
\item{x}{numeric vector of values}

\item{timestamps}{timestamp vector}

\item{apriori}{\link{apriori} class}

\item{...}{optional parameters, depending on signature:}

\item{plot}{prints comprehensive plots}

\item{verbose}{prints comprehensive information}

\item{title}{adds title to the plot}
}
\value{
Logical vector with same length as x, specifying TRUE for the duration of a temporal change.
}
\description{
This function marks temporalchanges in the input vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = numeric,apriori = apriori}: 
}}

