% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drifts.R
\name{Drift}
\alias{Drift}
\alias{Drift.logical}
\alias{Drift.Arima}
\title{Drift object}
\usage{
Drift(x, ...)

\method{Drift}{logical}(x, mu, timestamp, rate, significance)

\method{Drift}{Arima}(model, timestamps)
}
\description{
Drift object holds all the necessary information about the detected drift.
}
\keyword{internal}
