% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_reference.R
\name{drift_reference.aggregate}
\alias{drift_reference.aggregate}
\title{Aggregate df to specified scalefactor}
\usage{
drift_reference.aggregate(df, scalefactor.sec, is.reference = FALSE)
}
\description{
Aggregate df to specified scalefactor
}
\keyword{internal}
