% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_reference.R
\name{drift_reference.differentiate}
\alias{drift_reference.differentiate}
\title{Takes the differences between x and the reference values}
\usage{
drift_reference.differentiate(x, timestamps, reference)
}
\value{
A data.frame (x, timestamps, reference.id) with the differences.
Note that in case of more than 1 reference is supplied, there might be
duplicate timestamps based on different reference barometers.
}
\description{
Takes the differences between x and the reference values
}
\keyword{internal}
