% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{dtft}
\alias{dtft}
\title{Discrete-Time Fourier Transform (DTFT)}
\usage{
dtft(frequency, x, timestamps, plot = FALSE)
}
\description{
See here: https://en.wikipedia.org/wiki/Discrete-time_dtft
Note that there is also Discrete Fourier Transform O(n^2) (DFT), which is
equivalent to the better optimized Fast Fourier Transform O(nlog(n)) (FFT)
implementation. The DFT is derived from the DTFT as the minimal discrete
representation in the frequency domain of the original finite series. The
DTFT is more general since the frequency domain representation is continous.
Hence why this function will compute the corresponding complex number for
any required frequency.
}
\keyword{internal}
