% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fbasis}
\alias{fbasis}
\title{Fourier Basis}
\usage{
fbasis(timestamps, intercept = FALSE, trend = FALSE, frequencies = NULL)
}
\description{
If frequencies are not given, then default harmonics of Fourier Series
are used. Optionaly, one can choose to also add a trend and an intercept.
}
\keyword{internal}
