% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{gr_vectorsize.inch}
\alias{gr_vectorsize.inch}
\title{Computes the vector size in inch}
\usage{
gr_vectorsize.inch(x1, y1, x0 = 0, y0 = 0)
}
\arguments{
\item{x1, y1}{Endpoint of the vector.}

\item{x0, y0}{Origin of the vector. Default: (0,0).}
}
\description{
This function is mainly inspired by the following post:
https://stackoverflow.com/a/52690054/1548942
}
\keyword{internal}
