% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwloggeR.R
\docType{package}
\name{gwloggeR-package}
\alias{gwloggeR}
\alias{gwloggeR-package}
\title{gwloggeR: Validate and annotate raw logger air and hydrostatic pressure time series}
\description{
Groundwater divers measure hydrostatic pressure, air pressure and temperature. These raw time series contain issues such as outliers, drift,... This package provides functions to facilitate the validation workflow of groundwater related time series. Each of the functions takes a time series and annotates problematic observations with a controlled vocabulary of flags.
}
\details{
TODO: package details.
}
\references{
\href{https://github.com/DOV-Vlaanderen/groundwater-logger-validation/tree/master/gwloggeR}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dov-vlaanderen.github.io/groundwater-logger-validation/gwloggeR/docs/}
  \item \url{https://github.com/DOV-Vlaanderen/groundwater-logger-validation/gwloggeR}
  \item Report bugs at \url{https://github.com/DOV-Vlaanderen/groundwater-logger-validation/issues}
}

}
\author{
\strong{Maintainer}: Davor Josipovic \email{davor.josipovic@inbo.be} (\href{https://orcid.org/0000-0003-1204-8637}{ORCID})

}
\keyword{internal}
