% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelshifts.R
\name{Levelshifts}
\alias{Levelshifts}
\title{Levelshifts object}
\usage{
Levelshifts()
}
\description{
Levelshifts object holds all the necessary information about detected levelshifts.
}
\keyword{internal}
