% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_drifts.R
\name{model_drifts.fit}
\alias{model_drifts.fit}
\title{Fits a drift detection model}
\usage{
model_drifts.fit(dr.x, dr.ts, ar1, dfdiff)
}
\value{
Arima or ArimaExt object
}
\description{
Fits a drift detection model
}
\keyword{internal}
