% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{Optimizer}
\alias{Optimizer}
\title{Hydrostatic pressure model optimizer.}
\usage{
Optimizer(dx, types, indexes, mu, sigma2, par.init = NULL)
}
\value{
Returns an \link{Optimizer.Result} object.
}
\description{
Object for optimizing the hydrostatic pressure model. One needs to
initialize the object, and subsequently call the optimize() member-function.
}
\keyword{internal}
