% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{Optimizer.Result}
\alias{Optimizer.Result}
\title{Hydrostatic pressure model optimizer result}
\usage{
Optimizer.Result(logLval, par, types, types.significant, indexes, optim)
}
\description{
This is the object that is returned by the \link{Optimizer} object.
}
\keyword{internal}
