% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.base}
\alias{plot.base}
\title{Base functions for plotting hydrostatic pressure data}
\usage{
\method{plot}{base}(data)
}
\description{
Base functions for plotting hydrostatic pressure data
}
\keyword{internal}
