% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{seeker}
\alias{seeker}
\title{Selects best model based on degrees of freedom.}
\usage{
seeker(dx, mu, sigma2, outlier, types)
}
\description{
This function starts with a baseline model, and calls
\link{sweep} subsequenty to increase the model complexity.
Once a more complex model fails a likelihood ratio (LR) test, the function
stops and returns the most likely model.
}
\keyword{internal}
