% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{sweep}
\alias{sweep}
\title{Hydrostatic pressure model selection step}
\usage{
sweep(
  dx,
  base.types = NULL,
  base.indexes = NULL,
  base.par = NULL,
  sweep.indexes,
  types,
  mu,
  sigma2,
  pt
)
}
\description{
Based on base model events (and corresponding indexes),
this function evaluates all possible configurations of a new event, computes
the likelihoods using \link{Optimizer} and updates the \link{ProgressTable}.
}
\keyword{internal}
