% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalchanges.R
\name{Temporalchanges}
\alias{Temporalchanges}
\title{Temporalchanges object}
\usage{
Temporalchanges(vec)
}
\description{
Temporalchanges object holds all the necessary information about detected temporal changes.
}
\keyword{internal}
