% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.detect_function}
\alias{test.detect_function}
\title{Function for uniform testing of the detect_...() functions}
\usage{
test.detect_function(
  fun,
  ...,
  RESULT.PATH,
  ATTRIB.PATH = NULL,
  IMG.PATH = NULL
)
}
\description{
Function for uniform testing of the detect_...() functions
}
\keyword{internal}
