% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{Version}
\alias{Version}
\title{Version object}
\usage{
Version(version, url = NULL)
}
\description{
Version object holds all the necessary information about the
algorithm used to generate the output.
}
\keyword{internal}
