% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apriori.R
\name{Apriori}
\alias{Apriori}
\alias{apriori}
\title{A-priori data information class}
\usage{
Apriori(
  data_type = c("air pressure", "hydrostatic pressure"),
  units = c("cmH2O")
)

apriori(
  data_type = c("air pressure", "hydrostatic pressure"),
  units = c("cmH2O")
)
}
\arguments{
\item{data_type}{character, specifying type of data}

\item{units}{character, specifying units of data}
}
\value{
Apriori object.
}
\description{
This function creates an Apriori class based on input.
}
\examples{
# For example:
ap <- Apriori("air pressure", "cmH2O")

}
