% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drifts.R
\name{detect_drift}
\alias{detect_drift}
\alias{detect_drift,numeric-method}
\title{Detect drift}
\usage{
detect_drift(
  x,
  timestamps,
  reference = list(),
  apriori = Apriori("air pressure", units = "cmH2O"),
  ...,
  plot = FALSE,
  verbose = FALSE,
  title = NULL,
  alpha = 0.01
)

\S4method{detect_drift}{numeric}(x, timestamps, reference, apriori, plot, verbose, title, alpha)
}
\arguments{
\item{x}{numeric vector of values}

\item{timestamps}{timestamp vector}

\item{reference}{a list of list(x, timestamps) pairs that are to be used as
reference for drift detection. The more data is supplied here, the better
the results. Note also that the barometer data supplied here should be
compensated for altitude.}

\item{apriori}{\link{Apriori} class. Defaults to air pressure in cmH2O.}

\item{...}{optional parameters, depending on signature:}

\item{plot}{prints diagnostic plots. Consult the dedicated section below for more information.}

\item{verbose}{adds extra information as attributes to the result.}

\item{title}{adds title to the plot}

\item{alpha}{significance level at which to detect drift. Defaults to 0.01.}
}
\value{
Logical vector with same length as x, specifying TRUE for drifting observations.
}
\description{
This function marks a drift in the input vector.
}
\section{Functions}{
\itemize{
\item \code{detect_drift,numeric-method}: 
}}

\section{Diagnostic plots}{
Note that the DTFT and Year plots have the drift effect subtracted. In the Year plots
this results to a more pronounced seasonal pattern, and in the DTFT plot the
amplitude of low frequencies (high periods) is tuned down.
}

\references{
\href{https://dov-vlaanderen.github.io/groundwater-logger-validation/gwloggeR/docs/articles/Airpressure.html}{Air pressure vignette}
}
