% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drifts.R
\name{detect_drift}
\alias{detect_drift}
\alias{detect_drift,numeric-method}
\title{Detect drift}
\usage{
detect_drift(
  x,
  timestamps,
  reference = list(),
  apriori = Apriori("air pressure", units = "cmH2O"),
  ...,
  plot = FALSE,
  verbose = FALSE,
  title = NULL,
  alpha = 0.01
)

\S4method{detect_drift}{numeric}(x, timestamps, reference, apriori, plot, verbose, title, alpha)
}
\arguments{
\item{x}{numeric vector of values}

\item{timestamps}{timestamp vector}

\item{reference}{a list of list(x, timestamps) pairs that are to be used as
reference for drift detection. The more data is supplied here, the better
the results.}

\item{apriori}{\link{Apriori} class. Defaults to air pressure in cmH2O.}

\item{...}{optional parameters, depending on signature:}

\item{plot}{prints diagnostic plots. Consult the dedicated section below for more information.}

\item{verbose}{adds extra information as attributes to the result.}

\item{title}{adds title to the plot}

\item{alpha}{significance level at which to detect drift. Defaults to 0.01.}
}
\value{
Logical vector with same length as x, specifying TRUE for drifting observations.
}
\description{
This function marks a drift in the input vector.
}
\details{
\subsection{Diagnostic plots}{

Top plot shows the time differences (hours) between measurements. 12 hours is
signified by the red horizontal line. Note that the y-scale is logarithmic.

The middle left plot is the original series (black) overlaying the reference (darkred).
In case more than one reference is supplied, only their median is shown.

The bottom left plot is the difference between the original and the reference series
(or their median in case more than 1 reference is supplied) and is best suited for drift detection.
The vertical dotted line signifies the most likely drift start date.
Subsequently, three colors are used based on the significance: red (p < 0.001), orange and green (p > 0.01).

The right 4 smaller plots are DFTF and Year plots for the original series x (top) and differences (bottom).
They are mainly used for the visualization of the seasonality patterns. The DFTF plot is in days while the
yearly plot is in months.
Note that the DTFT and Year plots have the drift effect subtracted. In the Year plots
this results to a more pronounced seasonal pattern, and in the DTFT plot this results in
lower amplitude of low frequencies (high periods).
}
}
\section{Functions}{
\itemize{
\item \code{detect_drift,numeric-method}: 
}}

\references{
\href{https://dov-vlaanderen.github.io/groundwater-logger-validation/gwloggeR/docs/articles/Airpressure-Drift.html}{Air pressure drift vignette}
}
