% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\docType{methods}
\name{detect_outliers}
\alias{detect_outliers}
\alias{detect_outliers,numeric,missing-method}
\alias{detect_outliers,numeric,apriori-method}
\title{Detect outliers}
\usage{
detect_outliers(x, apriori, ..., plot = FALSE, verbose = FALSE,
  title = NULL, timestamps = NULL)

\S4method{detect_outliers}{numeric,missing}(x, plot, verbose, title,
  timestamps)

\S4method{detect_outliers}{numeric,apriori}(x, apriori, plot, verbose,
  title, timestamps)
}
\arguments{
\item{x}{numeric vector of values}

\item{apriori}{\link{apriori} class}

\item{...}{optional parameters, depending on signature:}

\item{plot}{prints comprehensive plots}

\item{verbose}{prints comprehensive information}

\item{title}{adds title to the plot}

\item{timestamps}{timestamp vector. Has no direct meaning, except aestehtical
for the scatter plot if plot = TRUE.}
}
\value{
Logical vector with same length as x, specifying TRUE for an outlier.
}
\description{
This function marks outliers in the input vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = numeric,apriori = missing}: Only considers x, without any \emph{a-priori} information.
A normal distribution is assumed with mean and variance estimated using
median and MAD as described in Leys, 2013.

\item \code{x = numeric,apriori = apriori}: Takes \emph{a-priori} information about x into consideration.
}}

\examples{
# In case of a vector:
x <- c(1:10, 100)
detect_outliers(x)

# In case of a dataframe, select the column:
detect_outliers(cars$dist)

\dontrun{
# Or use the tidyverse approach:
library(magrittr)
cars \%>\% dplyr::mutate("outlier" = detect_outliers(dist))
}

}
\references{
Leys, C. e.a., Detecting outliers, 2013.
}
