% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(x)
}
\arguments{
\item{x}{numeric vector of values}
}
\value{
logical vector with same length as x, specifying TRUE for outlier
}
\description{
This function marks outliers in the input vector.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
