% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(x)
}
\arguments{
\item{x}{numeric vector of values}
}
\value{
logical vector with same length as x, specifying TRUE for outlier
}
\description{
This function marks outliers in the input vector.
}
\examples{
# In case of a vector:
x <- c(1:10, 100)
detect_outliers(x)

# In case of a dataframe, select the column:
detect_outliers(cars$dist)

# Or use the tidyverse approach:
library(magrittr)
cars \%>\% dplyr::mutate("outlier" = detect_outliers(dist))

}
