% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.leys.R
\name{detect_outliers_leys}
\alias{detect_outliers_leys}
\title{Detects outliers}
\usage{
detect_outliers_leys(x, p.value = 5e-04, verbose = FALSE,
  M = median(x, na.rm = TRUE), MAD = mad(x, na.rm = TRUE))
}
\description{
This function is based on Leys, C. e.a., Detecting outliers, 2013.
}
