% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drifts.R
\name{Drift}
\alias{Drift}
\alias{Drift.logical}
\alias{Drift.Arima}
\alias{Drift.ArimaExt}
\title{Drift object}
\usage{
Drift(x, ...)

\method{Drift}{logical}(
  x,
  mu,
  timestamp = as.POSIXct(NA),
  rate = as.numeric(NA),
  year.seasonality = c(sine = as.numeric(NA), cosine = NA),
  significance = as.numeric(NA)
)

\method{Drift}{Arima}(model, timestamps, ...)

\method{Drift}{ArimaExt}(model, timestamps, sig.treshold)
}
\description{
Drift object holds all the necessary information about the detected drift.
}
\keyword{internal}
