% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydorpressure.R
\name{Events}
\alias{Events}
\title{Events object}
\usage{
Events(results, index.offsets)
}
\arguments{
\item{results}{a list of one or multiple Optimizer.Result objects.}
}
\description{
Events is a dataframe object that holds the indexes,
corresponding types and parameters.
}
