% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{Outliers}
\alias{Outliers}
\title{Outliers object}
\usage{
Outliers(x.rejects, x.mean, x.sd, sigma.reject, alpha, type, fun.density,
  cutpoints)
}
\description{
Outliers object holds all the necessary information about detected outiers.
}
\keyword{internal}
