% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalchanges.R
\name{Temporalchanges}
\alias{Temporalchanges}
\alias{Temporalchanges.logical}
\alias{Temporalchanges.Events}
\title{Temporalchanges object}
\usage{
Temporalchanges(x)

\method{Temporalchanges}{logical}(x)

\method{Temporalchanges}{Events}(events)
}
\description{
Temporalchanges object holds all the necessary information about detected temporal changes.
}
\keyword{internal}
